app/requirements.txt
soundcloud-dl
spotify-dl
python-dotenv
pathlib
selenium
mp3-tagger

